+++ ribs/emitters/_gradient_arborescence_emitter.py
@@ -54,8 +54,7 @@ class GradientArborescenceEmitter(EmitterBase):
         :doc:`/tutorials/tom_cruise_dqd`
 
     Args:
-        archive (ribs.archives.ArchiveBase): An archive to use when creating and
-            inserting solutions. For instance, this can be
+        archive (ribs.archives.ArchiveBase): Archive of solutions, e.g.,
             :class:`ribs.archives.GridArchive`.
         x0 (np.ndarray): Initial solution.
         sigma0 (float): Initial step size / standard deviation of the