+++ tests/visualize/sliding_boundaries_archive_heatmap_test.py
@@ -49,6 +49,16 @@ def sliding_archive_2d():
 
 
 @pytest.fixture(scope="module")
+def sliding_archive_2d_empty():
+    """Same as above but without solutions."""
+    archive = SlidingBoundariesArchive(solution_dim=2,
+                                       dims=[10, 20],
+                                       ranges=[(-1, 1), (-1, 1)],
+                                       seed=42)
+    return archive
+
+
+@pytest.fixture(scope="module")
 def sliding_archive_2d_long():
     """Same as above, but the measure space is longer in one direction."""
     archive = SlidingBoundariesArchive(solution_dim=2,
@@ -107,6 +117,19 @@ def test_2d_long_transpose(sliding_archive_2d_long):
 def test_limits(sliding_archive_2d):
     plt.figure(figsize=(8, 6))
     sliding_boundaries_archive_heatmap(sliding_archive_2d, vmin=-1.0, vmax=-0.5)
+
+
+@image_comparison(baseline_images=["limits_when_empty"],
+                  remove_text=False,
+                  extensions=["png"])
+def test_limits_when_empty(sliding_archive_2d_empty):
+    plt.figure(figsize=(8, 6))
+    sliding_boundaries_archive_heatmap(
+        sliding_archive_2d_empty,
+        # Intentionally don't provide vmin or vmax.
+        vmin=None,
+        vmax=None,
+    )
 
 
 @image_comparison(baseline_images=["listed_cmap"],